HEEDLS-439 Split CustomPromptsServiceTests into multiple files by prompt type
